1. update app info and render update; 2. add event; 3. update u8 log;

+++ Src/Engine/Pipeline.cpp
@@ -64,7 +64,7 @@ void IPipeline::create(vk::CDevice::CPtr vDevice, VkRenderPass vRenderPass, VkEx
     _createV();
 }
 
-void IPipeline::setImageNum(size_t vImageNum)
+void IPipeline::setImageNum(uint32_t vImageNum)
 {
     if (vImageNum == m_ImageNum) return;
     m_ImageNum = vImageNum;
